backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
google-auth==2.36.0
google-api-python-client==2.154.0
python-dotenv==1.0.1
